checking out previously working branch

+++ Controller/main.h
@@ -1,5 +1,5 @@
 
 #include "defines.h"
 #include <avr/io.h>
-extern uint16_t fData[NUMBER_OF_SENSORS];
+extern uint32_t fData[NUMBER_OF_SENSORS];
 int main();
